Fix set_verify(VERIFY_PEER) to work without a callback
It seems that unless you call the get_current_cert method like this, that
client certificates don't work.

Ironically this was the same bug as seen in IO::Socket::SSL which led to me
to writing this module.  Yet I shipped the first release with the same bug!

I'd say this is probably a bug in OpenSSL introduced at some point, and
almost no software tickles it by requesting client cert validation but not
installing a callback like this which fetches the actual certificate for
inspection.  But not really sure, haven't confirmed.