X509 - remove call to free()
For some reason this seems to frequently segfault; possibly if the Session
it is allocated to is freed first then it will be a double free.  However
this is just a hunch and for now a leak is less scary than a segfault.

Also include a set of 'bad' (actually very minimal, non-CN containing) set
of certificates, and rename the other tests for segfault-y-ness and don't
ship the extra ones.  Really the t/04-*.t should all become one script.